+++ routes/ft/utils.js
@@ -1,4 +1,9 @@
+const axios = require('../../utils/axios');
+const cheerio = require('cheerio');
+const Parser = require('rss-parser');
+
 const ProcessFeed = ($, link) => {
+    const title = $('h1').text();
     let content = $('div.story-container');
 
     // 处理封面图片
@@ -27,9 +32,57 @@ const ProcessFeed = ($, link) => {
     });
     content = content.html();
 
-    return { content, author };
+    return { content, author, title };
+};
+
+const getData = async ({ site = 'www', channel }) => {
+    const parser = new Parser();
+    let feed;
+
+    if (channel) {
+        let channel = channel.toLowerCase();
+        channel = channel.split('-').join('/');
+
+        try {
+            feed = await parser.parseURL(`http://${site}.ftchinese.com/rss/${channel}`);
+        } catch (error) {
+            return {
+                title: `FT 中文网 ${channel} 不存在`,
+                description: `FT 中文网 ${channel} 不存在`,
+            };
+        }
+    } else {
+        feed = await parser.parseURL(`http://${site}.ftchinese.com/rss/feed`);
+    }
+
+    const items = await Promise.all(
+        feed.items.splice(0, 10).map(async (item) => {
+            const response = await axios({
+                method: 'get',
+                url: `${item.link}?full=y&archive`,
+            });
+
+            const $ = cheerio.load(response.data);
+            const result = ProcessFeed($, item.link);
+            const single = {
+                title: result.title,
+                description: result.content,
+                author: result.author,
+                pubDate: item.pubDate,
+                link: item.link,
+            };
+            return Promise.resolve(single);
+        })
+    );
+
+    return {
+        title: feed.title,
+        link: feed.link,
+        description: feed.description,
+        item: items,
+    };
 };
 
 module.exports = {
-    ProcessFeed,
+    getData,
 };
